Fixed issue of Year not being correctly polled.
Check_Edit:
 Pushing general YTD now pushes to correct location
 Cleaned up console log and made log more reader friendly

Current Issues:
 Pushing Tax YTD is not finding the correct entry and therefore will not push any data

+++ lib/src/main/java/Gradle_Payroll/gui/Main_Menu.java
@@ -6,6 +6,7 @@ import java.awt.Dimension;
 import java.awt.Toolkit;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.util.Calendar;
 
 import javax.swing.JDesktopPane;
 import javax.swing.JFrame;
@@ -24,14 +25,12 @@ public class Main_Menu extends JFrame implements ActionListener{
 
 	static JDesktopPane MainMenu;
 	static int CheckNum;
-	static double year;
+	static int year;
 	
 	
 	
 
 	public Main_Menu() {
-		
-	
 		
 		super("ACI Payroll");
